+++ CHANGELOG.md
@@ -5,3 +5,9 @@
 ## 0.0.1+1
 
 - Removed the (windows, linux, mac, and web) directories from the example
+
+## 0.0.2
+
+- Added support for FormData (fields and files) logging in the request.
+- Added capability to display the type of the request body you're sending (Json, or FormData)
+- Other minor bug fixes